Fix bug where unescape_string_literal fails
If a output stream does not have a valid encoding, unescape_string_literal
would fail with a TypeError. This occurs in cases such as using subprocess
to redirect outputs, etc.

PIP uses subprocess to run the content of setup.py directly with
subprocess and eval. The output redirection in subprocess causes the
output stream has encoding of "None", resulting #8. This fixes it.